Adds frame seeking hotkeys.
Also, refactor the method for finding the right timestamp to start/end
at when a trim exists. The new method uses a binary search, so it should
be more efficient than a linear seek. I use this method to wrap around
in the frame seeking.

Tested using trimmed videos at different speeds, and everything works as
expected.

Updated Readme with new hotkeys